+++ src/main/java/derniertourencotedor/modele/algo/rechercheLocale/successeur/AlgoTourneRechercheSuccPremier.java
@@ -21,9 +21,9 @@ public class AlgoTourneRechercheSuccPremier extends AlgoTourneRechercheLocale
         Tourne tourne = new Tourne(villes);
         for (int i = 1; i < tourneCourante.tailleTourne()-2; i++)
         {
-            float distance1 =
+            double distance1 =
                     villes.get(i - 1).distanceAvecVille(villes.get(i)) + villes.get(i+1).distanceAvecVille(villes.get(i+2));
-            float distance2 =
+            double distance2 =
                     villes.get(i - 1).distanceAvecVille(villes.get(i+1)) + villes.get(i).distanceAvecVille(villes.get(i+2));
             if(distance1 > distance2) tourne.intervertirVilles(villes.get(i), villes.get(i+1));
         }
